dispatcher cleanup
