+++ ledger/ledger-api-common/src/main/scala/com/digitalasset/platform/server/services/testing/ReferenceTimeService.scala
@@ -12,7 +12,7 @@ import com.digitalasset.api.util.TimestampConversion._
 import com.digitalasset.grpc.adapter.ExecutionSequencerFactory
 import com.digitalasset.ledger.api.v1.testing.time_service.TimeServiceGrpc.TimeService
 import com.digitalasset.ledger.api.v1.testing.time_service._
-import com.digitalasset.platform.akkastreams.SignalDispatcher
+import com.digitalasset.platform.akkastreams.dispatcher.SignalDispatcher
 import com.digitalasset.platform.api.grpc.GrpcApiService
 import com.digitalasset.platform.common.util.DirectExecutionContext
 import com.digitalasset.platform.server.api.validation.FieldValidations
