+++ ledger/ledger-api-common/src/test/suite/scala/com/digitalasset/platform/akkastreams/DispatcherIT.scala
@@ -8,7 +8,8 @@ import java.util.concurrent.atomic.{AtomicInteger, AtomicReference}
 import akka.NotUsed
 import akka.stream.scaladsl.{Keep, Sink}
 import com.digitalasset.ledger.api.testing.utils.AkkaBeforeAndAfterAll
-import com.digitalasset.platform.akkastreams.SteppingMode.OneAfterAnother
+import com.digitalasset.platform.akkastreams.dispatcher.Dispatcher
+import com.digitalasset.platform.akkastreams.dispatcher.SteppingMode.OneAfterAnother
 import org.scalatest.concurrent.ScalaFutures
 import org.scalatest.time.{Milliseconds, Seconds, Span}
 import org.scalatest.{Matchers, WordSpec}
